+++ go/internal/features/todoFetch/repository/interface.go
@@ -1,6 +1,6 @@
 package repository
 
-import "github.com/teruyoshi/todoApp/internal/features/todoCreate/entity"
+import "github.com/teruyoshi/todoApp/internal/features/todos/entity"
 
 // TodoRepository provides methods to fetch todos.
 type TodoRepository interface {
