+++ go/internal/features/todos/repository/interface.go
@@ -1,6 +1,6 @@
 package repository
 
-import "github.com/teruyoshi/todoApp/internal/features/todoCreate/entity"
+import "github.com/teruyoshi/todoApp/internal/features/todos/entity"
 
 type TodoRepository interface {
 	Create(t entity.Todo) (entity.Todo, error)
